refactor: changes according to pr
