--- src/httpClient/httpClientFactory.ts
@@ -1,6 +1,0 @@
-import axios, { AxiosInstance, AxiosRequestConfig } from 'axios';
-
-export const httpClientFactory = (httpClientConfig: AxiosRequestConfig): AxiosInstance => {
-  const axiosClient = axios.create(httpClientConfig);
-  return axiosClient;
-};
