+++ src/index.ts
@@ -8,11 +8,6 @@ import { ExitCodes, EXIT_CODE, SERVICES } from './common/constants';
 import { getCli } from './cli/cli';
 import { ShutdownHandler } from './common/shutdownHandler';
 
-interface IError {
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  error: (...data: any[]) => void;
-}
-
 const exitProcess = (): void => {
   const exitCode = container.isRegistered(EXIT_CODE) ? container.resolve<number>(EXIT_CODE) : ExitCodes.GENERAL_ERROR;
   process.exit(exitCode);
@@ -23,9 +18,9 @@ void getCli()
     await cli.parseAsync(hideBin(process.argv));
   })
   .catch((error: Error) => {
-    const loggerOrConsole: IError = container.isRegistered(SERVICES.LOGGER) ? container.resolve<Logger>(SERVICES.LOGGER) : console;
-    loggerOrConsole.error('failed initializing the cli');
-    loggerOrConsole.error(error);
+    const errorLogger = container.isRegistered(SERVICES.LOGGER) ? container.resolve<Logger>(SERVICES.LOGGER).error : console.error;
+    errorLogger('failed initializing the cli');
+    errorLogger(error);
   })
   .finally(() => {
     if (!container.isRegistered(ShutdownHandler)) {
